code rearrangement and compoentization
